河西总装信息管理系统

+++ hxzz/src/main/java/com/hxzz/demo/entity/PersonnelManagement.java
@@ -50,7 +50,7 @@ public class PersonnelManagement implements Serializable {
     private Integer shouldArrive;
 
     @TableField("actualArrive")
-    private Integer actualArrive;
+    private Float actualArrive;
 
     private LocalDate date;
