+++ hxzz/src/main/java/com/hxzz/demo/service/PersonnelManagementService.java
@@ -31,13 +31,13 @@ public interface PersonnelManagementService extends IService<PersonnelManagement
     PersonnelManagement infoSum(LocalDate date1, LocalDate date2);
 
     void addIN(String name, Integer establishment, Integer actualNumber, Integer availableNumber, Float attendanceRate,
-               Integer shouldArrive, Integer actualArrive, LocalDate date);
+               Integer shouldArrive, Float actualArrive, LocalDate date);
 
     void delIN(Integer id);
 
     void changeIN(Integer id, String name, Integer establishment, Integer actualNumber, Integer availableNumber,
                   Float attendanceRate,
-                  Integer shouldArrive, Integer actualArrive, LocalDate date);
+                  Integer shouldArrive, Float actualArrive, LocalDate date);
 
     PageInfo<PersonnelManagement> findAll(Integer pageNum, Integer pageSize, LocalDate date1, LocalDate date2);
